pyproject.toml
[project]
name = "screenshot-organiser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.2.16",
    "python-dotenv>=1.1.0",
]
